Added some basic checks for function/source/sink in cli as well
