+++ pulsar-client-tools/src/main/java/org/apache/pulsar/admin/cli/CmdSources.java
@@ -370,6 +370,7 @@ public class CmdSources extends CmdBase {
             if (isBlank(sourceConfig.getArchive())) {
                 throw new ParameterException("Source archive not specfied");
             }
+            org.apache.pulsar.common.functions.Utils.inferMissingArguments(sourceConfig);
             if (!Utils.isFunctionPackageUrlSupported(sourceConfig.getArchive()) &&
                 !sourceConfig.getArchive().startsWith(Utils.BUILTIN)) {
                 if (!new File(sourceConfig.getArchive()).exists()) {
